fastapi_service/requirements.txt
fastapi
uvicorn
torch
numpy
pillow
ultralytics